+++ examples/msg-passing.jl
@@ -75,8 +75,8 @@ function train(; kws...)
         WithGraph(fg, GraphConv(fg, args.hidden_dim=>args.target_dim)),
     ) |> device
 
-    # ADAM optimizer
-    opt = ADAM(args.η)
+    # Adam optimizer
+    opt = Adam(args.η)
     
     # parameters
     ps = Flux.params(model)