+++ src/GeometricFlux.jl
@@ -17,7 +17,6 @@ using Flux: glorot_uniform, leakyrelu, GRUCell, @functor
 using Graphs
 using NNlib, NNlibCUDA
 using Optimisers
-using Zygote
 
 import Word2Vec: word2vec, wordvectors, get_vector