+++ test/runtests.jl
@@ -11,7 +11,6 @@ using LinearAlgebra
 using NNlib, NNlibCUDA
 using SparseArrays: SparseMatrixCSC
 using Statistics: mean
-using Zygote
 using Test
 
 cuda_tests = [
@@ -28,7 +27,6 @@ tests = [
     "layers/pool",
     "layers/graphlayers",
     "sampling",
-    "embedding/node2vec",
     "models",
 ]
 
@@ -36,6 +34,10 @@ if CUDA.functional()
     append!(tests, cuda_tests)
 else
     @warn "CUDA unavailable, not testing GPU support"
+end
+
+if !Sys.iswindows()
+    push!(tests, "embedding/node2vec")
 end
 
 @testset "GeometricFlux" begin