+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.12.4]
+
+- replace `ADAM` as `Adam`
+
+## [0.12.3]
+
+- update doc for graph network
+
+## [0.12.2]
+
+- replace `Zygote.ignore` as `ChainRulesCore.ignore_derivatives`
+
 ## [0.12.1]
 
 - remove @deprecate